+++ examples/resources/project/import.sh
@@ -1,1 +1,0 @@
-terraform import project.myproject myproj