+++ tfc-testing/sample.tf
@@ -9,7 +9,7 @@ terraform {
   required_providers {
     project = {
       source  = "jfrog/project"
-      version = "1.5.3"
+      version = "1.9.9"
     }
   }
 }
@@ -19,7 +19,7 @@ provider "project" {
   oidc_provider_name = "terraform-cloud"
 }
 
-resource "project" "myproject" {
+resource "project_project" "myproject" {
   key = "myproj"
   display_name = "My Project"
   description  = "My Project"